refactor StepPreview into a component
