+++ src/webapp/components/markdown-editor/StepPreview.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { MarkdownViewer } from "../markdown-viewer/MarkdownViewer";
+import styled from "styled-components";
+import { ModalBody } from "../modal";
+
+export const StepPreview: React.FC<{
+    className?: string;
+    value?: string;
+}> = ({ className, value }) => {
+    if (!value) return null;
+
+    return (
+        <StyledModalBody className={className}>
+            <MarkdownViewer source={value} />
+        </StyledModalBody>
+    );
+};
+
+const StyledModalBody = styled(ModalBody)`
+    max-width: 600px;
+`;
